Fixed issues, added hooks, updated components
