+++ src/components/Authentication/index.jsx
@@ -12,7 +12,6 @@ export function Auth() {
       setLoading(true);
       const { error } = await supabase.auth.signInWithOtp({ email });
       if (error) throw error;
-      alert("Check your email for the login link!");
     } catch (error) {
       alert(error.error_description || error.message);
     } finally {
