+++ src/index.css
@@ -15,7 +15,7 @@
 }
 
 #root {
-  max-width: 600px;
+  max-width: 800px;
   margin: 0 auto;
   padding: 2rem;
   text-align: center;
